# openapi/openapi.yaml
openapi: "3.0.3"

info:
  title: NotaryLink API
  version: "1.0.0"
  description: API contract for the NotaryLink notarial platform

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Auth
    description: Authentication endpoints
  - name: User
    description: User profile endpoints

paths:
  /api/auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Authenticate an existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/verify-otp:
    post:
      tags: [Auth]
      operationId: verifyOtp
      summary: Verify the 6-digit OTP code sent by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
      responses:
        '200':
          description: OTP verified successfully
        '400':
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/resend-otp:
    post:
      tags: [Auth]
      operationId: resendOtp
      summary: Resend the OTP code to the user's email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendOtpRequest'
      responses:
        '200':
          description: OTP resent successfully
        '400':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/user/me:
    get:
      tags: [User]
      operationId: getCurrentUser
      summary: Get the currently authenticated user's profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    VerifyOtpRequest:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          minLength: 6
          maxLength: 6

    ResendOtpRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email

    AuthResponse:
      type: object
      required: [accessToken, tokenType]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          default: Bearer

    User:
      type: object
      required: [id, email, firstName, lastName]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [USER, NOTARY, ADMIN]
        otpVerified:
          type: boolean

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string